+++ packages/flutter_riverpod/lib/src/consumer.dart
@@ -296,7 +296,7 @@ typedef ConsumerBuilder = Widget Function(
 ///               builder: (BuildContext context, WidgetRef ref, Widget? child) {
 ///                 // This builder will only get called when the counterProvider
 ///                 // is updated.
-///                 final count = ref.watch(counterProvider.state).state;
+///                 final count = ref.watch(counterProvider);
 ///
 ///                 return Row(
 ///                   mainAxisAlignment: MainAxisAlignment.spaceEvenly,
@@ -316,7 +316,7 @@ typedef ConsumerBuilder = Widget Function(
 ///       ),
 ///       floatingActionButton: FloatingActionButton(
 ///         child: Icon(Icons.plus_one),
-///         onPressed: () => ref.read(counterProvider.state).state++,
+///         onPressed: () => ref.read(counterProvider.notifier).state++,
 ///       ),
 ///     );
 ///   }