+++ packages/riverpod/test/framework/provider_observer_test.dart
@@ -38,7 +38,7 @@ void main() {
         final container = createContainer(observers: [observer, observer2]);
         final dep = StateProvider((ref) => 0);
         final provider = Provider((ref) {
-          if (ref.watch(dep.state).state == 0) {
+          if (ref.watch(dep) == 0) {
             throw UnimplementedError();
           }
           return 0;
@@ -49,7 +49,7 @@ void main() {
         clearInteractions(observer);
         clearInteractions(observer2);
 
-        container.read(dep.state).state++;
+        container.read(dep.notifier).state++;
         await container.pump();
 
         verifyInOrder([
@@ -323,7 +323,7 @@ void main() {
         final container = createContainer(observers: [observer, observer2]);
         final dep = StateProvider((ref) => 0);
         final provider = Provider((ref) {
-          if (ref.watch(dep.state).state != 0) {
+          if (ref.watch(dep) != 0) {
             throw UnimplementedError();
           }
           return 0;
@@ -334,7 +334,7 @@ void main() {
         clearInteractions(observer);
         clearInteractions(observer2);
 
-        container.read(dep.state).state++;
+        container.read(dep.notifier).state++;
         await container.pump();
 
         verifyInOrder([