+++ packages/riverpod/test/providers/future_provider/auto_dispose_future_provider_test.dart
@@ -85,8 +85,7 @@ void main() {
       'when going from AsyncLoading to AsyncLoading, does not notify listeners',
       () async {
     final dep = StateProvider((ref) => Future.value(42));
-    final provider =
-        FutureProvider.autoDispose((ref) => ref.watch(dep.state).state);
+    final provider = FutureProvider.autoDispose((ref) => ref.watch(dep));
     final container = createContainer();
     final listener = Listener<AsyncValue<int>>();
 
@@ -102,7 +101,7 @@ void main() {
     );
 
     final completer = Completer<int>();
-    container.read(dep.state).state = completer.future;
+    container.read(dep.notifier).state = completer.future;
 
     container.listen(provider, listener, fireImmediately: true);
 
@@ -115,7 +114,7 @@ void main() {
       ),
     );
 
-    container.read(dep.state).state = Future.value(21);
+    container.read(dep.notifier).state = Future.value(21);
 
     verifyNoMoreInteractions(listener);
 
@@ -179,7 +178,7 @@ void main() {
 
     verifyOnly(listener, listener(null, const AsyncValue.loading()));
 
-    container.read(dep.state).state++;
+    container.read(dep.notifier).state++;
     await container.pump();
 
     verifyNoMoreInteractions(listener);
@@ -201,7 +200,7 @@ void main() {
 
     verifyOnly(listener, listener(any, any));
 
-    container.read(dep.state).state++;
+    container.read(dep.notifier).state++;
     await container.pump();
 
     verifyNoMoreInteractions(listener);