+++ packages/riverpod/test/providers/provider/auto_dispose_provider_test.dart
@@ -50,7 +50,7 @@ void main() {
         final container = createContainer();
         Object? err;
         final provider = Provider.autoDispose<int>((ref) {
-          if (ref.watch(dep.state).state) {
+          if (ref.watch(dep)) {
             try {
               ref.state;
             } catch (e) {
@@ -63,7 +63,7 @@ void main() {
         container.read(provider);
         expect(err, isNull);
 
-        container.read(dep.state).state = true;
+        container.read(dep.notifier).state = true;
         container.read(provider);
 
         expect(err, isStateError);