+++ packages/riverpod/test/providers/stream_provider/auto_dispose_stream_provider_test.dart
@@ -70,8 +70,7 @@ void main() {
         'when going from AsyncLoading to AsyncLoading, does not notify listeners',
         () async {
       final dep = StateProvider((ref) => Stream.value(42));
-      final provider =
-          StreamProvider.autoDispose((ref) => ref.watch(dep.state).state);
+      final provider = StreamProvider.autoDispose((ref) => ref.watch(dep));
       final container = createContainer();
       final listener = Listener<AsyncValue<int>>();
       final controller = StreamController<int>();
@@ -88,7 +87,7 @@ void main() {
         const AsyncData<int>(42),
       );
 
-      container.read(dep.state).state = controller.stream;
+      container.read(dep.notifier).state = controller.stream;
       container.listen(provider, listener, fireImmediately: true);
 
       verifyOnly(
@@ -100,7 +99,7 @@ void main() {
         ),
       );
 
-      container.read(dep.state).state = Stream.value(21);
+      container.read(dep.notifier).state = Stream.value(21);
 
       verifyNoMoreInteractions(listener);
 
@@ -152,7 +151,7 @@ void main() {
 
       verifyOnly(listener, listener(null, const AsyncValue.loading()));
 
-      container.read(dep.state).state++;
+      container.read(dep.notifier).state++;
       await container.pump();
 
       verifyNoMoreInteractions(listener);
@@ -173,7 +172,7 @@ void main() {
 
       verifyOnly(listener, listener(any, any));
 
-      container.read(dep.state).state++;
+      container.read(dep.notifier).state++;
       await container.pump();
 
       verifyNoMoreInteractions(listener);
@@ -194,7 +193,7 @@ void main() {
 
       verifyOnly(listener, listener(any, any));
 
-      container.read(dep.state).state++;
+      container.read(dep.notifier).state++;
       await container.pump();
 
       verifyNoMoreInteractions(listener);